pyproject.toml
[build-system]
# we use the hatch build system
# https://hatch.pypa.io/latest/config/build/
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pySMWSync"
authors = [
  {name = "Wolfgang Fahl", email = "wf@bitplan.com"}
]
maintainers = [
  { name = "Wolfgang Fahl", email = "wf@bitplan.com" },
]
readme = "README.md"
license= "Apache-2.0"
dependencies = [
	 # https://pypi.org/project/py-3rdparty-mediawiki/
     'py-3rdparty-mediawiki>=0.9.1',
     # https://github.com/WolfgangFahl/pyMetaModel
     'pyMetaModel>=0.2.3',
     # https://pypi.org/project/colorama/
     'colorama'
]

requires-python = ">=3.8"
classifiers=[
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License"
    ]
dynamic = ["version", "description"]
[tool.hatch.version]
path = "smwsync/__init__.py"

[project.urls]
Home = "https://github.com/WolfgangFahl/pySMWSync"
Documentation = "https://wiki.bitplan.com/index.php/pySMWSync"
Source = "https://github.com/WolfgangFahl/pySMWSync"

[project.optional-dependencies]
test = [
  "green",
]

[tool.hatch.build.targets.wheel]
packages = [
  "smwsync",
]

[project.scripts]
smwsync = "smwsync.synccmd:main"
semslides = "slides.semslides:main"
